backend/requirements.txt
beautifulsoup4
dj-rest-auth
django
django-autoslug
django-cors-headers
django-environ
djangorestframework
djangorestframework-simplejwt
gunicorn
pillow
pokemontcgsdk
requests
